refactor listcommand and predicates
