+++ src/main/java/seedu/address/model/person/predicates/FixedAddressPredicate.java
@@ -1,4 +1,7 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
+
+import seedu.address.model.person.Address;
+import seedu.address.model.person.ReadOnlyPerson;
 
 import java.util.function.Predicate;
