+++ src/main/java/seedu/address/model/person/predicates/ShowSpecifiedPersonPredicate.java
@@ -1,4 +1,6 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicates;
+
+import seedu.address.model.person.ReadOnlyPerson;
 
 import java.util.function.Predicate;
