+++ src/main/java/seedu/address/model/person/predicates/UniqueAttributePredicate.java
@@ -1,0 +1,6 @@
+package seedu.address.model.person.predicates;
+
+import java.util.function.Predicate;
+
+public interface UniqueAttributePredicate<T> extends Predicate<T> {
+}
